[DEV] Fix webpack for development workflow (#1163)
