--- web/src/hooks/AssetPath.ts
@@ -1,7 +1,0 @@
-import { useBasePath } from "./BasePath";
-
-__webpack_public_path__ = "/"
-
-if (useBasePath() !== "") {
-  __webpack_public_path__ = useBasePath() + "/"
-}
