--- web/src/hooks/BasePath.ts
@@ -1,5 +1,0 @@
-import { useEmbeddedVariable } from "./Configuration";
-
-export function useBasePath() {
-    return useEmbeddedVariable("basepath");
-}
