+++ web/src/utils/AssetPath.ts
@@ -1,0 +1,7 @@
+import { getBasePath } from "./BasePath";
+
+__webpack_public_path__ = "/"
+
+if (getBasePath() !== "") {
+  __webpack_public_path__ = getBasePath() + "/"
+}
