+++ web/src/utils/BasePath.ts
@@ -1,0 +1,5 @@
+import { getEmbeddedVariable } from "./Configuration";
+
+export function getBasePath() {
+    return getEmbeddedVariable("basepath");
+}
